+++ CHANGELOG.md
@@ -12,6 +12,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Requests resulting in `503 Service Unavailable` errors are now retried.
   Attempting to download a document before translation is completed will now
   wait and retry (up to 5 times by default), rather than throwing an exception.
+### Fixed
+* Use `Locale.ENGLISH` when changing string case.
+  * Thanks to [seratch](https://github.com/seratch).
 
 
 ## [0.1.3] - 2022-09-09